Bound array indices, reject implausible clock writes, stop value truncation
Three unguarded array indices, all reachable without an attacker:

  webUIupdates.cpp indexed HC_PROGRAM[][9] with hc1_program/hc2_program
  and BURNER_STATE[][5] with kmStatus.BurnerStates - raw fields decoded
  straight from the serial bus. km271.cpp already clamps the same fields
  with limit() when it publishes them, but the web path did not, so one
  corrupted telegram (or a burner variant reporting an unexpected value)
  hands ArduinoJson a wild pointer. The bad value is latched into the
  status copy, so it would then repeat on every web UI page load.

  config.lang, config.mqtt.lang and config.log.filter are used as array
  indices too. The web callbacks validate them, but configLoadFromFile()
  does not, so a hand-edited or uploaded config.json could put an
  arbitrary value there. Clamped centrally in configFinalCheck(), which
  runs after every load, and cfg_logger_filter now clamps at the callback
  like cfg_lang already did.

km271SetDateTimeDTI()/km271SetDateTimeNTP() now refuse to write an
implausible date to the boiler. The automatic caller in main.cpp already
gates on getLocalTime(), but the cmd/datetime MQTT topic calls
km271SetDateTimeNTP() directly - so an HA automation firing right after a
reboot, before NTP resolved, could set the Logamatic's clock to 1970. The
web UI path builds its tm from separate date and time messages and could
submit a partially-filled struct. Since the Logamatic's day/night heating
program runs off this clock, a bogus value silently changes when the
house is heated.

km271sendCmd() took int8_t, so the narrowing happened before the range
check and the check validated the truncated value: publishing 286 to a
30..60 °C setpoint wrapped to 30 and was accepted as valid, with a
confirmation message reporting 30. Widened to int, so the existing range
checks see what was actually sent and reject it. The MQTT payload
variable was int16_t for the same reason - it also capped the oil counter
(in 1/100 L) at 327.67 L before wrapping negative - and is now long,
parsed with atol().

+++ src/mqtt.cpp
@@ -328,10 +328,14 @@ void processMqttMessage() {
   ESP_LOGD(TAG, "process msg from buffer: %s, %s", msgCpy.topic, msgCpy.payload);
 
   // payload as number
-  int16_t intVal = 0;
+  // wide enough that the range checks inside km271sendCmd() see the value the
+  // user actually sent: as int16_t, publishing 286 to a 30..60 setpoint wrapped
+  // to 30 and was accepted as valid, and the oil counter (in 1/100 L) could only
+  // express 0..327.67 L before wrapping negative
+  long intVal = 0;
   float floatVal = 0.0;
   if (msgCpy.len > 0) {
-    intVal = atoi(msgCpy.payload);
+    intVal = atol(msgCpy.payload);
     floatVal = atoff(msgCpy.payload);
   }
